+++ epan/dissectors/packet-tcp.c
@@ -718,7 +718,7 @@ finished_fwd:
 		 * duplicate ack
 		 * then this is a fast retransmission
 		 */
-		t=(pinfo->fd->abs_ts.secs-tcpd->rev->lastacktime.secs)*1000000000;
+		t=(guint32) (pinfo->fd->abs_ts.secs-tcpd->rev->lastacktime.secs)*1000000000;
 		t=t+(pinfo->fd->abs_ts.nsecs)-tcpd->rev->lastacktime.nsecs;
 		if( tcpd->rev->dupacknum>=2
 		&&  tcpd->rev->lastack==seq
@@ -734,7 +734,7 @@ finished_fwd:
 		 * seen sequence number, then it is an OUT-OF-ORDER segment.
 		 *   (3ms is an arbitrary number)
 		 */
-		t=(pinfo->fd->abs_ts.secs-tcpd->fwd->nextseqtime.secs)*1000000000;
+		t=(guint32) (pinfo->fd->abs_ts.secs-tcpd->fwd->nextseqtime.secs)*1000000000;
 		t=t+(pinfo->fd->abs_ts.nsecs)-tcpd->fwd->nextseqtime.nsecs;
 		if( t<3000000 ){
 			if(!tcpd->ta){
